+++ app/client/styles/props/color/blue.ts
@@ -15,20 +15,6 @@ export const blueValues = {
 };
 
 export const blue = {
-  '50': 'var(--color-blue-50)',
-  '100': 'var(--color-blue-100)',
-  '200': 'var(--color-blue-200)',
-  '300': 'var(--color-blue-300)',
-  '400': 'var(--color-blue-400)',
-  '500': 'var(--color-blue-500)',
-  '600': 'var(--color-blue-600)',
-  '700': 'var(--color-blue-700)',
-  '800': 'var(--color-blue-800)',
-  '900': 'var(--color-blue-900)',
-  '950': 'var(--color-blue-950)',
-};
-
-export const blueTags = {
   '50': css`var(--color-blue-50)`,
   '100': css`var(--color-blue-100)`,
   '200': css`var(--color-blue-200)`,
