+++ app/client/styles/props/color/green.ts
@@ -15,20 +15,6 @@ export const greenValues = {
 };
 
 export const green = {
-  '50': 'var(--color-green-50)',
-  '100': 'var(--color-green-100)',
-  '200': 'var(--color-green-200)',
-  '300': 'var(--color-green-300)',
-  '400': 'var(--color-green-400)',
-  '500': 'var(--color-green-500)',
-  '600': 'var(--color-green-600)',
-  '700': 'var(--color-green-700)',
-  '800': 'var(--color-green-800)',
-  '900': 'var(--color-green-900)',
-  '950': 'var(--color-green-950)',
-};
-
-export const greenTags = {
   '50': css`var(--color-green-50)`,
   '100': css`var(--color-green-100)`,
   '200': css`var(--color-green-200)`,
