+++ app/client/styles/props/color/index.ts
@@ -1,8 +1,8 @@
-export { red, redValues, redTags } from './red';
-export { orange, orangeValues, orangeTags } from './orange';
-export { green, greenValues, greenTags } from './green';
-export { blue, blueValues, blueTags } from './blue';
-export { yellow, yellowValues, yellowTags } from './yellow';
-export { indigo, indigoValues, indigoTags } from './indigo';
-export { neutral, neutralValues, neutralTags } from './neutral';
-export { violet, violetValues, violetTags } from './violet';
+export { red, redValues } from './red';
+export { orange, orangeValues } from './orange';
+export { green, greenValues } from './green';
+export { blue, blueValues } from './blue';
+export { yellow, yellowValues } from './yellow';
+export { indigo, indigoValues } from './indigo';
+export { neutral, neutralValues } from './neutral';
+export { violet, violetValues } from './violet';
