+++ app/client/styles/props/color/indigo.ts
@@ -15,20 +15,6 @@ export const indigoValues = {
 };
 
 export const indigo = {
-  '50': 'var(--color-indigo-50)',
-  '100': 'var(--color-indigo-100)',
-  '200': 'var(--color-indigo-200)',
-  '300': 'var(--color-indigo-300)',
-  '400': 'var(--color-indigo-400)',
-  '500': 'var(--color-indigo-500)',
-  '600': 'var(--color-indigo-600)',
-  '700': 'var(--color-indigo-700)',
-  '800': 'var(--color-indigo-800)',
-  '900': 'var(--color-indigo-900)',
-  '950': 'var(--color-indigo-950)',
-};
-
-export const indigoTags = {
   '50': css`var(--color-indigo-50)`,
   '100': css`var(--color-indigo-100)`,
   '200': css`var(--color-indigo-200)`,
