+++ app/client/styles/props/color/neutral.ts
@@ -15,20 +15,6 @@ export const neutralValues = {
 };
 
 export const neutral = {
-  '50': 'var(--color-neutral-50)',
-  '100': 'var(--color-neutral-100)',
-  '200': 'var(--color-neutral-200)',
-  '300': 'var(--color-neutral-300)',
-  '400': 'var(--color-neutral-400)',
-  '500': 'var(--color-neutral-500)',
-  '600': 'var(--color-neutral-600)',
-  '700': 'var(--color-neutral-700)',
-  '800': 'var(--color-neutral-800)',
-  '900': 'var(--color-neutral-900)',
-  '950': 'var(--color-neutral-950)',
-};
-
-export const neutralTags = {
   '50': css`var(--color-neutral-50)`,
   '100': css`var(--color-neutral-100)`,
   '200': css`var(--color-neutral-200)`,
