+++ app/client/styles/props/color/orange.ts
@@ -15,20 +15,6 @@ export const orangeValues = {
 };
 
 export const orange = {
-  '50': 'var(--color-orange-50)',
-  '100': 'var(--color-orange-100)',
-  '200': 'var(--color-orange-200)',
-  '300': 'var(--color-orange-300)',
-  '400': 'var(--color-orange-400)',
-  '500': 'var(--color-orange-500)',
-  '600': 'var(--color-orange-600)',
-  '700': 'var(--color-orange-700)',
-  '800': 'var(--color-orange-800)',
-  '900': 'var(--color-orange-900)',
-  '950': 'var(--color-orange-950)',
-};
-
-export const orangeTags = {
   '50': css`var(--color-orange-50)`,
   '100': css`var(--color-orange-100)`,
   '200': css`var(--color-orange-200)`,
