+++ app/client/styles/props/color/violet.ts
@@ -15,20 +15,6 @@ export const violetValues = {
 };
 
 export const violet = {
-  '50': 'var(--color-violet-50)',
-  '100': 'var(--color-violet-100)',
-  '200': 'var(--color-violet-200)',
-  '300': 'var(--color-violet-300)',
-  '400': 'var(--color-violet-400)',
-  '500': 'var(--color-violet-500)',
-  '600': 'var(--color-violet-600)',
-  '700': 'var(--color-violet-700)',
-  '800': 'var(--color-violet-800)',
-  '900': 'var(--color-violet-900)',
-  '950': 'var(--color-violet-950)',
-};
-
-export const violetTags = {
   '50': css`var(--color-violet-50)`,
   '100': css`var(--color-violet-100)`,
   '200': css`var(--color-violet-200)`,
