+++ app/client/styles/props/index.ts
@@ -1,10 +1,6 @@
 export * as color from './color';
-export { space, spaceValues, spaceTags } from './space';
-export {
-  letterSpacing,
-  letterSpacingValues,
-  letterSpacingTags,
-} from './letterSpacing';
-export { radius, radiusValues, radiusTags } from './radius';
-export { font, fontValues, fontTags } from './font';
-export { lineHeight, lineHeightValues, lineHeightTags } from './lineHeight';
+export { space, spaceValues } from './space';
+export { letterSpacing, letterSpacingValues } from './letterSpacing';
+export { radius, radiusValues } from './radius';
+export { font, fontValues } from './font';
+export { lineHeight, lineHeightValues } from './lineHeight';
