+++ app/client/styles/props/letterSpacing.ts
@@ -9,14 +9,6 @@ export const letterSpacingValues = {
 };
 
 export const letterSpacing = {
-  tighter: 'var(--letter-spacing-tighter)',
-  tight: 'var(--letter-spacing-tight)',
-  normal: 'var(--letter-spacing-normal)',
-  loose: 'var(--letter-spacing-loose)',
-  looser: 'var(--letter-spacing-looser)',
-};
-
-export const letterSpacingTags = {
   tighter: css`var(--letter-spacing-tighter)`,
   tight: css`var(--letter-spacing-tight)`,
   normal: css`var(--letter-spacing-normal)`,
