+++ app/client/styles/props/space.ts
@@ -14,19 +14,6 @@ export const spaceValues = {
 };
 
 export const space = {
-  '3xs': 'var(--space-3xs)',
-  '2xs': 'var(--space-2xs)',
-  xs: 'var(--space-xs)',
-  sm: 'var(--space-sm)',
-  md: 'var(--space-md)',
-  lg: 'var(--space-lg)',
-  xl: 'var(--space-xl)',
-  '2xl': 'var(--space-2xl)',
-  '3xl': 'var(--space-3xl)',
-  '4xl': 'var(--space-4xl)',
-};
-
-export const spaceTags = {
   '3xs': css`var(--space-3xs)`,
   '2xs': css`var(--space-2xs)`,
   xs: css`var(--space-xs)`,
