+++ design-system/index.ts
@@ -1,0 +1,4 @@
+import { run } from './main/run';
+
+const args = process.argv;
+run(args).catch((err) => console.error(err));
