+++ design-system/utils/stringUtils.ts
@@ -1,0 +1,3 @@
+export function toCamelCase(str: string): string {
+  return str.replace(/-([a-z])/g, (g) => g[1].toUpperCase());
+}
